WebSockets/tests/submissions/Opera/interfaces/WebSocket/url/002.html

<!doctype html>
<title>WebSockets: setting url</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function() {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  ws.url = SCHEME_AND_DOMAIN+'/test';
  assert_equals(ws.url, SCHEME_AND_DOMAIN+'/');
});
</script>
